+++ src/main/kotlin/dev/avt/api/admin/absentie/readRequestedHours/ReadRequestedHoursResponse.kt
@@ -1,6 +1,5 @@
 package dev.avt.api.admin.absentie.readRequestedHours
 
-import dev.avt.api.admin.absentie.approveHours.HoursRequestedDataFormat
 import kotlinx.serialization.Serializable
 
 @Serializable
